Add new Text storybook doc
